requirements.txt
asgiref==3.3.1
Django==3.1.6
django-extensions==3.1.0
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
Faker==5.8.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
